Improve structure, add registrations in apps.py as per new django specification. Upgrade views, serializers for newer versions of django and drf. Add new settings to support separation of email and in app notifications (is_read flag) and wrappable kronos tasks for job control.
