+++ action_notifications/apiurls.py
@@ -1,4 +1,5 @@
-from django.conf.urls import patterns, include, url
+from django.conf.urls import include, url
+from django.urls import re_path
 
 from rest_framework.routers import DefaultRouter
 
@@ -7,6 +8,6 @@ from action_notifications import views
 router = DefaultRouter()
 router.register(r'action_notification', views.ActionNotificationViewSet)
 
-urlpatterns = patterns('',
-    url(r'^v1/', include(router.urls)),
+urlpatterns = (
+    re_path(r'^v1/', include(router.urls)),
 )
